+++ zarrs_storage/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - Bump `unsafe_cell_slice` to 0.2.0
  - **Breaking**: Change `ByteRange::FromEnd` to `ByteRange::Suffix`
+ - **Breaking**: implement `AsyncReadableWritableStorageTraits` for `T: AsyncReadableStorageTraits + AsyncWritableStorageTraits`
 
 ### Removed
  - **Breaking**: Remove `ByteRange::offset()`